backend/requirements.txt
flask
flask-cors
torch
torchvision
timm
pillow
google-generativeai
reportlab
python-dotenv
python-dateutil
fpdf2